refactor: Separate GameStatusUpdater from GameSaleTaskScheduler
- @Transactional self-invocation 문제 해결
- GameStatusUpdater로 상태 변경 트랜잭션 로직 분리
- GameSaleTaskScheduler는 스케줄 등록 역할만 담당
- GameStatusUpdater 단위/통합 테스트 추가
- GameSaleTaskScheduler 테스트 등록 검증으로 축소
- NotificationEvent 대응
